Fixes issue #5 (off-by-one error in 'feed()').
'http_parser_execute()' does not mark the last byte of headers as consumed
after calling the 'on_headers_complete()' callback if 'http_parser_pause()'
is called from the callback.

This fix forces the parser to consume the extra byte by unpausing, then calling
'http_parser_execute()' again with the last byte.  Note that this byte is
always available since it had to be there to trigger the callback that pauses
the parser.