Allow calling `skip_body()` before feeding any header data.
If `mySkipBody` is reset in the `on_message_begin()`, then a single `feed()`
call containing all the headers will not allow a call `skip_body()` before the
parser enters the body parsing state.  This fix ensures that `skip_body()` can
be called before the first call to `feed()` for a given response.